refactor(runner): improve code organization
- Move ssh.ts to firecracker/guest.ts (groups with related code)
- Rename tar.ts to utils.ts and remove dead code (createScriptsTarBuffer)
- Update imports in executor.ts

The guest.ts rename reflects purpose (VM guest communication) rather
than implementation detail (SSH protocol). Dead code removal follows
YAGNI principle - createScriptsTarBuffer was copied from E2B but never
used since scripts are uploaded individually via SSH.
