+++ turbo/apps/runner/src/lib/firecracker/index.ts
@@ -7,3 +7,4 @@
 export * from "./client.js";
 export * from "./network.js";
 export * from "./vm.js";
+export * from "./guest.js";
